Prevent circular imports in ts code (closes: #523)
