+++ js/compiler.ts
@@ -2,7 +2,9 @@
 import * as ts from "typescript";
 import { assetSourceCode } from "./assets";
 import * as deno from "./deno";
-import { libdeno, window, globalEval } from "./globals";
+import { globalEval } from "./global-eval";
+import { libdeno } from "./libdeno";
+import { window } from "./globals";
 import * as os from "./os";
 import { RawSourceMap } from "./types";
 import { assert, log, notImplemented } from "./util";
