+++ js/deno.ts
@@ -1,4 +1,4 @@
 // Copyright 2018 the Deno authors. All rights reserved. MIT license.
 // Public deno module.
 export { exit, readFileSync, writeFileSync } from "./os";
-export { libdeno } from "./globals";
+export { libdeno } from "./libdeno";
